requirement.txt
Flask
Flask-Session
pandas
open_clip_torch
deep-translator
pillow
opencv-python
lancedb
torch
scenedetect[opencv]